// model/skillcategory.type.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * describes a skill category that references skills
 */
export interface SkillCategorySchema {
    skill_category_id?: string;
    display_name: string;
    skills?: string[];
    color?: string;
    index?: number;
    [k: string]: unknown;
}